refactor: place Next cache handler adapter
